+++ server/pub/api.ts
@@ -2,18 +2,9 @@ import { z } from 'zod';
 import { Request, Express } from 'express';
 import { extendZodWithOpenApi } from '@anatine/zod-openapi';
 import { initServer } from '@ts-rest/express';
-import mime from 'mime-types';
-import multer from 'multer';
 
 import * as types from 'types';
-import {
-	type ImportCreatePubParams,
-	type ImportMethod,
-	type MetadataOptions,
-	type ProposedMetadata,
-} from 'utils/api/schemas/pub';
 
-import { createPubAttribution } from 'server/pubAttribution/queries';
 import { BadRequestError, ForbiddenError, NotFoundError } from 'server/utils/errors';
 import { getInitialData } from 'server/utils/initData';
 import { indexByProperty } from 'utils/arrays';
@@ -31,21 +22,14 @@ import { getPubDraftDoc } from 'server/utils/firebaseAdmin';
 import { writeDocumentToPubDraft } from 'server/utils/firebaseTools';
 import { isDuqDuq, isProd } from 'utils/environment';
 import { ensureUserIsCommunityAdmin } from 'utils/ensureUserIsCommunityAdmin';
-import { getTmpDirectoryPath } from 'workers/tasks/import/tmpDirectory';
-import { importFiles } from 'workers/tasks/import/import';
-import { BaseSourceFile } from 'utils/api/schemas/import';
-import { labelFiles } from 'client/containers/Pub/PubDocument/PubFileImport/formats';
 import { omitKeys } from 'utils/objects';
-import { uploadAndConvertImages } from 'utils/import/uploadAndConvertImages';
 import { Pub } from './model';
-import {
-	canCreatePub,
-	canDestroyPub,
-	getUpdatablePubFields,
-	managerUpdatableFields,
-} from './permissions';
+
+import { canCreatePub, canDestroyPub, getUpdatablePubFields } from './permissions';
+
 import { createPub, destroyPub, findPub, importToPub, updatePub } from './queries';
 import { getPubsById, queryPubIds } from './queryMany';
+import { convertFiles, createUploadMiddleware, handleImport } from './import';
 
 extendZodWithOpenApi(z);
 
@@ -105,278 +89,6 @@ const getRequestIds = createGetRequestIds<{
 	pubId?: string;
 	createPubToken?: string | null;
 }>();
-
-/**
- * We need to dynamically create the middleware in order to set a different temporary directory for each request, otherwise there's a chance that two requests will try to write to the same file.
- */
-const createUploadMiddleware = async () => {
-	const tmpDirPath = await getTmpDirectoryPath();
-
-	const storage = multer.diskStorage({
-		destination: (req, file, cb) => {
-			// @ts-expect-error we are mutating baby
-			req.tmpDir = tmpDirPath;
-			cb(null, tmpDirPath);
-		},
-		filename: (req, file, cb) => {
-			// Instead of using file.originalname, pull from req.body
-			// This assumes that filenames are provided in pairs with files
-
-			const f = req.body.filenames;
-
-			// if only one filename is provided, req.body.filenames will be a string
-			const suppliedFilename = Array.isArray(f) ? f.shift() : f;
-
-			const filename = suppliedFilename ?? file.originalname;
-
-			if (!filename) {
-				throw new BadRequestError(
-					new Error(
-						'No filename provided, either in the body or as the original filename.\n Check the order of your files and filenames, filenames should come before the files they refer to.',
-					),
-				);
-			}
-
-			if (file.mimetype === 'application/octet-stream') {
-				const mimeType = mime.contentType(filename);
-				if (mimeType) {
-					file.mimetype = mimeType;
-				}
-			}
-
-			cb(null, suppliedFilename || file.originalname);
-		},
-	});
-
-	const upload = multer({ storage }).array('files');
-	return upload;
-};
-
-const convertFiles = async ({ files, tmpDir }: ConvertInput) => {
-	const sourceFilesFromNormalFiles = (files as Express.Multer.File[]).map(
-		(file): BaseSourceFile & { tmpPath: string } => ({
-			clientPath: file.filename ?? file.originalname,
-			tmpPath: file.path,
-			state: 'complete',
-			loaded: file.size,
-			total: file.size,
-		}),
-	);
-
-	const uploadedAndConvertedFiles = await uploadAndConvertImages(
-		sourceFilesFromNormalFiles,
-		tmpDir,
-	);
-	const labeledFiles = labelFiles(uploadedAndConvertedFiles);
-
-	return importFiles({
-		sourceFiles: labeledFiles,
-		importerFlags: {},
-		tmpDirPath: tmpDir,
-	});
-};
-
-function addAttributionsToPub({
-	pubId,
-	attributions,
-	existingAttributions = [],
-	shouldMatchUsers = true,
-}: {
-	pubId: string;
-	attributions: { name: string; mostLikelyUserId?: string }[];
-	existingAttributions?: types.PubAttribution[];
-	shouldMatchUsers?: boolean;
-}) {
-	return Promise.all(
-		attributions
-			.filter(
-				(attr) =>
-					!existingAttributions.some(
-						(existingAttr) =>
-							shouldMatchUsers && existingAttr.userId === attr.mostLikelyUserId,
-					),
-			)
-			.map(async (attribution, idx) => {
-				const { name, mostLikelyUserId } = attribution;
-				const pubAttribution = await createPubAttribution({
-					pubId,
-					order: 1 / 2 ** idx,
-					...(shouldMatchUsers && mostLikelyUserId
-						? { userId: mostLikelyUserId }
-						: {
-								name,
-						  }),
-				});
-
-				return pubAttribution as types.PubAttribution;
-			}),
-	);
-}
-
-const getDetectedParamsToUpdate = ({
-	proposedMetadata,
-	pubCreateParams,
-	overrides,
-	overrideAll,
-}: {
-	proposedMetadata: ProposedMetadata;
-	pubCreateParams?: ImportCreatePubParams;
-	overrides?: MetadataOptions['overrides'];
-	overrideAll?: MetadataOptions['overrideAll'];
-}) => {
-	const detectedMetadata = omitKeys(proposedMetadata, ['attributions', 'metadata']);
-
-	const fieldToOverrideWithDetected = Array.isArray(overrides) ? overrides : [overrides];
-
-	const newParams = Object.entries(detectedMetadata).reduce((acc, [key, val]) => {
-		// if we're creating a pub we always want to use the proposed metadata
-		// if the user hasn't specified a value for this field
-		if (pubCreateParams && !(key in acc)) {
-			acc[key] = val;
-			return acc;
-		}
-
-		// this value has been specified by the user, and we do want to override it with the detected value
-		if (overrideAll || fieldToOverrideWithDetected?.some((field) => field === key)) {
-			acc[key] = val;
-		}
-
-		return acc;
-	}, pubCreateParams ?? {});
-
-	return newParams;
-};
-
-const maybeAddAttributionsToPub = async ({
-	pubId,
-	proposedAttributions,
-	attributions,
-}: {
-	pubId: string;
-	proposedAttributions?: ProposedMetadata['attributions'];
-	attributions?: boolean | 'match';
-}) => {
-	if (attributions === false || !proposedAttributions?.length) {
-		return [];
-	}
-
-	const addedAttributions = addAttributionsToPub({
-		pubId,
-		attributions: proposedAttributions.map(({ name, users }) => ({
-			name,
-			mostLikelyUserId: users?.[0]?.id,
-		})),
-		shouldMatchUsers: attributions === 'match',
-	});
-
-	return addedAttributions;
-};
-
-type ConvertInput = {
-	userId?: string;
-	tmpDir: string;
-	files: Express.Multer.File[];
-};
-
-type ToPubImportInput = ConvertInput & {
-	pubId: string;
-	metadataOptions?: MetadataOptions;
-	method?: ImportMethod;
-};
-type CreatePubImportInput = ConvertInput & {
-	tmpDir: string;
-	files: Express.Multer.File[];
-	communityId: string;
-	pubCreateParams: ImportCreatePubParams;
-	metadataOptions?: MetadataOptions;
-};
-
-function handleImport(options: ToPubImportInput): Promise<{
-	status: 200;
-	body: {
-		doc: types.DocJson;
-		pub: types.DefinitelyHas<types.Pub, 'attributions'>;
-	};
-}>;
-function handleImport(options: CreatePubImportInput): Promise<{
-	status: 201;
-	body: { doc: types.DocJson; pub: types.DefinitelyHas<types.Pub, 'attributions'> };
-}>;
-async function handleImport(options: ToPubImportInput | CreatePubImportInput) {
-	const { files, tmpDir, metadataOptions, userId } = options;
-
-	if ('pubId' in options) {
-		const { pubId, method } = options;
-
-		const pub = await Pub.findOne({
-			where: { id: pubId },
-		});
-
-		if (!pub) {
-			throw new NotFoundError();
-		}
-
-		const { proposedMetadata, doc } = await convertFiles({ files, tmpDir });
-
-		const newParams = getDetectedParamsToUpdate({
-			proposedMetadata,
-			overrides: metadataOptions?.overrides,
-			overrideAll: metadataOptions?.overrideAll,
-		});
-
-		const modifiedText = await writeDocumentToPubDraft(pub.id, doc, { method });
-
-		const updatedAttributes = await updatePub(
-			{ ...newParams, pubId },
-			managerUpdatableFields,
-			userId,
-		);
-		const updatedAttributions = await maybeAddAttributionsToPub({
-			pubId,
-			proposedAttributions: proposedMetadata.attributions,
-			attributions: metadataOptions?.attributions,
-		});
-
-		const outputPub = Object.assign(pub.toJSON(), updatedAttributes, {
-			attributions: updatedAttributions ?? [],
-		}) as types.DefinitelyHas<types.Pub, 'attributions'>;
-
-		return { status: 200, body: { doc: modifiedText, pub: outputPub } };
-	}
-
-	const { communityId } = options;
-
-	const { proposedMetadata, doc } = await convertFiles({ files, tmpDir });
-
-	const newParams = getDetectedParamsToUpdate({
-		proposedMetadata,
-		pubCreateParams: options.pubCreateParams,
-		overrides: metadataOptions?.overrides,
-		overrideAll: metadataOptions?.overrideAll,
-	});
-
-	const pub = await createPub(
-		{
-			communityId,
-			...newParams,
-		},
-		userId,
-	);
-
-	await writeDocumentToPubDraft(pub.id, doc);
-
-	const addedAttributions = await maybeAddAttributionsToPub({
-		pubId: pub.id,
-		proposedAttributions: proposedMetadata.attributions,
-		attributions: metadataOptions?.attributions,
-	});
-
-	const outputPub = Object.assign(pub.toJSON(), {
-		attributions: addedAttributions ?? [],
-	}) as types.DefinitelyHas<types.Pub, 'attributions'>;
-
-	return { status: 201, body: { doc, pub: outputPub } };
-}
 
 const s = initServer();
